refactor: Begins the rename process
